Update package.json, CI workflow & refactor the structure of the project
